# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cupslock API Documentation",
        "description": "Dokumentasi API untuk sistem Cupslock, meliputi autentikasi dan fitur utama aplikasi.",
        "contact": {
            "name": "Cupslock Dev Team",
            "email": "developer@cupslock.test"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register user baru",
                "description": "Mendaftarkan user baru dengan data name, username, email, dan password.",
                "operationId": "33da49124bf72d9a48fdde8ff88f7578",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "username",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Andika Wiguna"
                                    },
                                    "username": {
                                        "type": "string",
                                        "example": "andika"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "andika@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "secret"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Register success"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login user",
                "description": "Login menggunakan email dan password untuk mendapatkan token.",
                "operationId": "68c5cab6a386991c65ac59f24f9695a2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "andika@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "secret"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login success"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout user",
                "description": "Menghapus token aktif untuk user yang sedang login.",
                "operationId": "a03f1137c4c903d1965e492c020f7ff6",
                "responses": {
                    "200": {
                        "description": "Logout success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/user": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Lihat user login",
                "description": "Mengambil data user yang sedang login menggunakan token Sanctum.",
                "operationId": "833e72f98ff5afd7cc6f8952f80cd809",
                "responses": {
                    "200": {
                        "description": "Berhasil menampilkan data user"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Menampilkan daftar produk",
                "operationId": "7cd9d394f726672223f216b77acab086",
                "responses": {
                    "200": {
                        "description": "Daftar produk berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Cappuccino"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Kopi dengan susu dan foam"
                                            },
                                            "price": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 25
                                            },
                                            "image": {
                                                "type": "string",
                                                "example": "https://example.com/image.jpg"
                                            },
                                            "best_seller": {
                                                "type": "boolean",
                                                "example": true
                                            },
                                            "category_id": {
                                                "type": "integer",
                                                "example": 2
                                            },
                                            "category_name": {
                                                "type": "string",
                                                "example": "Coffee"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Tambah produk baru",
                "operationId": "e83af3dd12f9ebc4ad755304893c016f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "category_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Espresso"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Kopi hitam pekat tanpa gula"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 18
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": "https://example.com/image.jpg"
                                    },
                                    "best_seller": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Produk berhasil ditambahkan"
                    }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Menampilkan detail produk berdasarkan ID",
                "operationId": "fa07b5c9bbd4d51f1f3763ca56e0908b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail produk berhasil diambil"
                    },
                    "404": {
                        "description": "Produk tidak ditemukan"
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update data produk",
                "operationId": "e0e4bcc68686fa05279e23580e41bea1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Latte"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Kopi dengan susu lembut"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 25
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": "https://example.com/new-image.jpg"
                                    },
                                    "best_seller": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produk berhasil diperbarui"
                    },
                    "404": {
                        "description": "Produk tidak ditemukan"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Hapus produk berdasarkan ID",
                "operationId": "bae80a1b7d88ec2b8f025a048aa00f86",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produk berhasil dihapus"
                    },
                    "404": {
                        "description": "Produk tidak ditemukan"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        },
        {
            "name": "Products",
            "description": "Products"
        }
    ],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}